requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
boto3==1.24.36
botocore==1.27.36
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-environ==0.9.0
django-redis==5.2.0
django-render-partial==0.4
django-storages==1.12.3
django-taggit==3.0.0
idna==3.3
jmespath==1.0.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
urllib3==1.26.10
wrapt==1.14.1
